found the root of the zombie shell issue
On termination the bash shell object would clean it self up so it could
be reused with a fresh init. However, shells that had been started with
a trap e.g. like MTM-SSH, would run into an issue where the base shell
would terminate on ssh exiting, but because bash cleaned itself up, the
child could attempt sending more commands, which spawned a new base
shell, however this new base shell would not have the guard process, as
it terminated with the first base shell.

If the new shell was not shut down cleanly, e.g. php timeout, runtime
exception whatever the new base shell would keep on living for 1000 days
(limit of the sleep that keeps the stdIn pipe open).

The change adds a boolean class attribute to termination, when set the
shell cannot be reused it is forever dead.

In addition the termination logic got a make over to deal with various
edge cases where a terminating shell was in a dirty state e.g. the user
issued an "exit" command outside the terminate() function. The shell
will now exit clean even if the state is not as expected.

+++ Src/Models/Shells/Bash/Termination.php
@@ -33,10 +33,11 @@ class Termination extends \MTM\Shells\Models\Shells\Base
 		if ($this->getParent() !== null) {
 			return $this->getParent()->isBaseTerm();
 		} else {
-			$exists	= $this->getBasePipes()->getStdOut()->getExists();
-			if ($exists === true) {
-				$exists	= $this->getBasePipes()->getStdIn()->getExists();
-				if ($exists === true) {
+			
+			if ($this->_isTerm === false && $this->_termActive === false) {
+				//expensive, this is a shell call
+				$isRun		= \MTM\Utilities\Factories::getSoftware()->getOsTool()->pidRunning($this->_spawnPid);
+				if ($isRun === true) {
 					return false;
 				}
 			}
@@ -45,49 +46,68 @@ class Termination extends \MTM\Shells\Models\Shells\Base
 	}
 	public function terminate()
 	{
-		if ($this->_isInit === true) {
-			$this->_isInit	= null;
-			if (is_object($this->getChild()) === true) {
-				$this->getChild()->terminate();
-			}
-			if ($this->isBaseTerm() === false) {
+		if ($this->_isTerm === false && $this->_termActive === false) {
+			$this->_termActive	= true;
+		
+			if ($this->_isInit === true) {
 				
-				if ($this->getParent() !== null) {
+				if (is_object($this->getChild()) === true) {
+					$this->getChild()->terminate();
+				}
+				if ($this->getParent() === null) {
+					
+					//check if the pid is active. e.g MTM-SSH sets up a trap that terminates the shell on exit
+					//if the shell is no longer active, the std pipes are gone
+					$isRun		= \MTM\Utilities\Factories::getSoftware()->getOsTool()->pidRunning($this->_spawnPid);
+					if ($isRun === true) {
+						//make sure the last command is dead, give it one sec to exit
+						//we really dont care if it exits clean or not we are the base shell
+						//and need to shut down. if we wait too long any read error will not be thrown
+						$this->issueSigInt(false, 1000);
+						
+						//exit the shell
+						$cmdObj		= $this->getCmd();
+						$strCmd		= "exit";
+						$regEx		= false;
+						$timeout	= 0;
+						if ($this->getParent() !== null) {
+							$regEx		= "(".preg_quote($this->getParent()->getRegEx()).")";
+							$timeout	= $cmdObj->getTimeout();
+						}
+						$cmdObj->setCmd($strCmd)->setDelimitor($regEx)->setTimeout($timeout);
+						$cmdObj->get(false);
+						
+						//the watcher process is looking for the presense of procLock
+						//this is the emergency breake if everything else fails
+						//process will be dead within 10 sec
+						$this->getPipes()->getLock()->delete();
+					}
+					$this->_basePipes	= null;
+					
+				} elseif ($this->isBaseTerm() === false) {
 					//make sure the last command is dead, give it the default amount of time
 					//we really do need to get the prompt of our parent shell will have trouble
 					$this->issueSigInt(false);
-				} else {
-					//make sure the last command is dead, give it one sec to exit
-					//we really dont care if it exits clean or not we are the base shell
-					//and need to shut down. if we wait too long any read error will not be thrown
-					$this->issueSigInt(false, 1000);
-				}
-	
-				//exit the shell
-				$cmdObj		= $this->getCmd();
-				$strCmd		= "exit";
-				$regEx		= false;
-				$timeout	= 0;
-				if ($this->getParent() !== null) {
-					$regEx		= "(".preg_quote($this->getParent()->getRegEx()).")";
-					$timeout	= $cmdObj->getTimeout();
-				}
-				$cmdObj->setCmd($strCmd)->setDelimitor($regEx)->setTimeout($timeout);
-				$cmdObj->get(false);
-	
-				if ($this->getParent() === null) {
-					//the watcher process is looking for the presense of procLock
-					//this is the emergency breake if everything else fails
-					//process will be dead within 10 sec
-					$this->getPipes()->getLock()->delete();
-					$this->_basePipes	= null;
-				} else {
+					
+					//exit the shell
+					$cmdObj		= $this->getCmd();
+					$strCmd		= "exit";
+					$regEx		= false;
+					$timeout	= 0;
+					if ($this->getParent() !== null) {
+						$regEx		= "(".preg_quote($this->getParent()->getRegEx()).")";
+						$timeout	= $cmdObj->getTimeout();
+					}
+					$cmdObj->setCmd($strCmd)->setDelimitor($regEx)->setTimeout($timeout);
+					$cmdObj->get(false);
+					
 					$this->getParent()->setChild(null);
 					$this->setParent(null);
+					
 				}
 			}
-			
-			$this->_isInit	= false;
+			$this->_isTerm		= true;
+			$this->_termActive	= false;
 		}
 	}
 }
